cleanup data-space temp developments
