+++ .changeset/early-chairs-join.md
@@ -1,0 +1,38 @@
+---
+'@finos/legend-application-pure-ide-deployment': patch
+'@finos/legend-application-taxonomy-deployment': patch
+'@finos/legend-application-taxonomy-bootstrap': patch
+'@finos/legend-application-studio-deployment': patch
+'@finos/legend-extension-store-service-store': patch
+'@finos/legend-application-query-deployment': patch
+'@finos/legend-application-studio-bootstrap': patch
+'@finos/legend-extension-format-json-schema': patch
+'@finos/legend-application-query-bootstrap': patch
+'@finos/legend-extension-store-relational': patch
+'@finos/legend-extension-dsl-persistence': patch
+'@finos/legend-extension-store-flat-data': patch
+'@finos/legend-extension-dsl-data-space': patch
+'@finos/legend-extension-format-graphql': patch
+'@finos/legend-extension-format-morphir': patch
+'@finos/legend-extension-tracer-zipkin': patch
+'@finos/legend-extension-dsl-diagram': patch
+'@finos/legend-extension-dsl-mastery': patch
+'@finos/legend-extension-dsl-service': patch
+'@finos/legend-application-pure-ide': patch
+'@finos/legend-application-taxonomy': patch
+'@finos/legend-application-studio': patch
+'@finos/legend-extension-dsl-text': patch
+'@finos/legend-application-query': patch
+'@finos/legend-query-builder': patch
+'@finos/legend-server-depot': patch
+'@finos/legend-application': patch
+'@finos/legend-server-sdlc': patch
+'@finos/legend-dev-utils': patch
+'@finos/stylelint-config-legend-studio': patch
+'@finos/legend-storage': patch
+'@finos/eslint-plugin-legend-studio': patch
+'@finos/legend-shared': patch
+'@finos/babel-preset-legend-studio': patch
+'@finos/legend-graph': patch
+'@finos/legend-art': patch
+---
