+++ packages/legend-application-taxonomy/src/application/LegendTaxonomyNavigation.ts
@@ -28,7 +28,7 @@ export const LEGEND_TAXONOMY_ROUTE_PATTERN = Object.freeze({
   EXPLORE_TAXONOMY_TREE: `/tree/:${LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.TAXONOMY_TREE_KEY}`,
   EXPLORE_TAXONOMY_TREE_NODE: `/tree/:${LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.TAXONOMY_TREE_KEY}/:${LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.TAXONOMY_PATH}`,
   EXPLORE_TAXONOMY_TREE_NODE_DATA_SPACE: `/tree/:${LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.TAXONOMY_TREE_KEY}/:${LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.TAXONOMY_PATH}/:${LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.GAV}/:${LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.DATA_SPACE_PATH}`,
-  STANDALONE_DATA_SPACE_VIEWER: `/dataspace/:${LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.GAV}/:${LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.DATA_SPACE_PATH}`,
+  VIEW_DATA_SPACE: `/dataspace/:${LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.GAV}/:${LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.DATA_SPACE_PATH}`,
 });
 
 export type LegendTaxonomyPathParams = {
@@ -38,7 +38,7 @@ export type LegendTaxonomyPathParams = {
   [LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.DATA_SPACE_PATH]?: string;
 };
 
-export type LegendTaxonomyStandaloneDataSpaceViewerPathParams = {
+export type DataSpacePreviewPathParams = {
   [LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.GAV]: string;
   [LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.DATA_SPACE_PATH]: string;
 };
@@ -75,11 +75,11 @@ export const generateExploreTaxonomyTreeNodeDataSpaceRoute = (
     },
   );
 
-export const generateStandaloneDataSpaceViewerRoute = (
+export const generateDataSpaceViewerRoute = (
   GAVCoordinates: string,
   dataSpacePath: string,
 ): string =>
-  generatePath(LEGEND_TAXONOMY_ROUTE_PATTERN.STANDALONE_DATA_SPACE_VIEWER, {
+  generatePath(LEGEND_TAXONOMY_ROUTE_PATTERN.VIEW_DATA_SPACE, {
     [LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.GAV]: GAVCoordinates,
     [LEGEND_TAXONOMY_ROUTE_PATTERN_TOKEN.DATA_SPACE_PATH]: dataSpacePath,
   });
