+++ packages/legend-application-taxonomy/src/stores/data-space-preview/DataSpacePreviewStore.ts
@@ -36,24 +36,25 @@ import {
   assertErrorThrown,
 } from '@finos/legend-shared';
 import { makeObservable, flow, observable, flowResult } from 'mobx';
-import type { LegendTaxonomyPluginManager } from '../application/LegendTaxonomyPluginManager.js';
-import type { LegendTaxonomyApplicationStore } from './LegendTaxonomyBaseStore.js';
+import type { LegendTaxonomyPluginManager } from '../../application/LegendTaxonomyPluginManager.js';
+import type { LegendTaxonomyApplicationStore } from '../LegendTaxonomyBaseStore.js';
 import {
   EXTERNAL_APPLICATION_NAVIGATION__generateDataSpaceQueryEditorUrl,
-  type LegendTaxonomyStandaloneDataSpaceViewerPathParams,
-} from '../application/LegendTaxonomyNavigation.js';
+  type DataSpacePreviewPathParams,
+} from '../../application/LegendTaxonomyNavigation.js';
 import {
   createViewProjectHandler,
   createViewSDLCProjectHandler,
-} from './LegendTaxonomyDataSpaceViewerHelper.js';
+} from '../LegendTaxonomyDataSpaceViewerHelper.js';
 
-export class StandaloneDataSpaceViewerStore {
+export class DataSpacePreviewStore {
   readonly applicationStore: LegendTaxonomyApplicationStore;
   readonly depotServerClient: DepotServerClient;
   readonly graphManagerState: BasicGraphManagerState;
   readonly pluginManager: LegendTaxonomyPluginManager;
 
   readonly initState = ActionState.create();
+
   viewerState?: DataSpaceViewerState | undefined;
 
   constructor(
@@ -74,15 +75,15 @@ export class StandaloneDataSpaceViewerStore {
     this.pluginManager = applicationStore.pluginManager;
   }
 
-  *initialize(
-    params: LegendTaxonomyStandaloneDataSpaceViewerPathParams,
-  ): GeneratorFn<void> {
+  *initialize(params: DataSpacePreviewPathParams): GeneratorFn<void> {
     this.initState.inProgress();
     this.initState.setMessage(`Initializing...`);
+
     try {
       const { gav, dataSpacePath } = params;
       const { groupId, artifactId, versionId } = parseGAVCoordinates(gav);
 
+      // initialize
       yield this.graphManagerState.graphManager.initialize(
         {
           env: this.applicationStore.config.env,
@@ -105,6 +106,7 @@ export class StandaloneDataSpaceViewerStore {
           this.depotServerClient.getProject(groupId, artifactId),
         )) as PlainObject<ProjectData>,
       );
+
       // analyze data space
       const analysisResult = (yield DSL_DataSpace_getGraphManagerExtension(
         this.graphManagerState.graphManager,
