+++ packages/legend-application-taxonomy/style/index.scss
@@ -15,6 +15,6 @@
  */
 
 @forward 'deps';
-@forward 'standalone-data-space-viewer';
 @forward 'taxonomy-node-viewer';
 @forward 'taxonomy-explorer';
+@forward 'data-space-preview';
