+++ packages/legend-extension-dsl-data-space/src/stores/studio/DataSpacePreviewState.ts
@@ -132,11 +132,6 @@ export class DataSpacePreviewState extends EditorExtensionState {
             );
           },
         },
-        {
-          HACKY__previewExperimentalFeatures:
-            this.editorStore.applicationStore.config.options
-              .HACKY__previewExperimentalFeatures,
-        },
       );
       this.loadDataSpaceState.pass();
     } catch (error) {
