+++ packages/legend-extension-dsl-data-space/style/_deps.scss
@@ -14,4 +14,7 @@
  * limitations under the License.
  */
 
+@import url('@ag-grid-community/styles/ag-grid.css');
+@import url('@ag-grid-community/styles/ag-theme-balham.css');
+
 @import url('@finos/legend-extension-dsl-diagram/lib/index.css');
